Remove deprecated function insertRule(), replace with appendRule() https://bugs.webkit.org/show_bug.cgi?id=150188
Unreviewed. This was a mistake in my review of Ryan's patch.
The CSSOM is confusing. This test was not always using a CSSKeyframesRule
object, so it should still be calling insertRule. This does produce
a console message, so added that to the expected result.

* http/tests/css/resources/shared-stylesheet-mutation.js:
(executeTests): Back to insertRule.
* http/tests/css/shared-stylesheet-mutation-expected.txt:
* http/tests/css/shared-stylesheet-mutation-preconstruct-expected.txt:

* http/tests/css/resources/shared-stylesheet-mutation.js:
(executeTests): Back to insertRule.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@191158 268f45cc-cd09-0410-ab3c-d52691b4dbfc