Null dereference loading Blink layout test fast/forms/color/input-color-onchange-event.html https://bugs.webkit.org/show_bug.cgi?id=150192 <rdar://problem/23135050>
Reviewed by Darin Adler.

Source/WebCore:

Calling internals.selectColorInColorChooser() with a non-Element would
cause a null dereference. This is because in such case, the implementation
method is passed a null pointer and we would fail to null-check it. This
patch now does the null-check.

No new tests, existing test was updated.

* testing/Internals.cpp:
(WebCore::Internals::selectColorInColorChooser):

LayoutTests:

Import more checks from the corresponding Blink test.

* fast/forms/color/input-color-onchange-event-expected.txt:
* fast/forms/color/input-color-onchange-event.html:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@191294 268f45cc-cd09-0410-ab3c-d52691b4dbfc