ASSERTION FAILED: m_state == Initialized in SubresourceLoader::didReceiveResponse() https://bugs.webkit.org/show_bug.cgi?id=150327
Reviewed by Antti Koivisto.

This is how it happens:

1. print() is called while the document is still loading, so
   m_shouldPrintWhenFinishedLoading is set to true
2. DataURLDecoder::decode() finishes in the work queue thread,
   the completion handler is scheduled in the main thread
3. The load is cancelled
  3.1. SubresourceLoader::willCancel sets m_state = Finishing
  3.2. DOMWindow::finishedLoading() is called, and since
       m_shouldPrintWhenFinishedLoading is true, it does the print.
  3.3. Cancellation finishes and ResourceLoader::releaseResources()
       is called that sets m_reachedTerminalState = true

So, between 3.1 and 3.3, the state is Finishing, but
m_reachedTerminalState is false. What happens, in the GTK+ port at
least, is that the nested main loop used to make print()
synchronous, processes the DataURLDecoder::decode() completion
handler that was pending. The completion handler returns early if
m_reachedTerminalState is true, but it's not yet in this
particular case. So, it ends up calling didReceiveResponse,
because the decode didn't fail, when the subresource loader state
is Finishing.

I think there are two things here. One is that we shouldn't start
a print that was waiting for the load to finish when it
failed. That would fix the problem. But it's probably a good idea
to also check for cancellation in the DataURLDecoder::decode()
completion handler.

Fixes printing/print-close-crash.html in GTK+ Debug.

* loader/ResourceLoader.cpp:
(WebCore::ResourceLoader::loadDataURL): Return early from
DataURLDecoder::decode() completion handler if the load was cancelled.
* page/DOMWindow.cpp:
(WebCore::DOMWindow::finishedLoading): Do not start a print that
was witing for the load to finish when it failed.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@191332 268f45cc-cd09-0410-ab3c-d52691b4dbfc