Source/WebCore: [css-grid] Fix availableLogicalSpace computation with non-zero baseSize flex tracks https://bugs.webkit.org/show_bug.cgi?id=150359
Reviewed by Zalan Bujtas.

The availableLogicalSpace computation was incorrect whenever
the flex tracks had a non-zero baseSize before the 1fr unit
size resolution. That happened because when assigning the new
baseSize to the flex track, we were unconditionally
subtracting the whole baseSize to the
availableLogicalSpace. That's correct if the track is a "pure"
flex track, i.e. 2fr, but if the track had a non-zero baseSize
(like minmax(10px, 1fr)) then both the new and the old base
sizes were incorrectly used to compute the
availableLogicalSpace.

We can test the amount of remaining freeSpace by using content
distribution to align and item place on a non-zero baseSize
flex track. The content distribution will be accurate if and
only if the availableLogicalSpace computation is correct.

Test: fast/css-grid-layout/flex-content-distribution.html

* rendering/RenderGrid.cpp:
(WebCore::RenderGrid::computeUsedBreadthOfGridTracks):

LayoutTests:
[css-grid] Fix freeSpace computation with non-zero baseSize flex tracks
https://bugs.webkit.org/show_bug.cgi?id=150359

Reviewed by Zalan Bujtas.

* fast/css-grid-layout/flex-content-distribution-expected.txt: Added.
* fast/css-grid-layout/flex-content-distribution.html: Added.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@191385 268f45cc-cd09-0410-ab3c-d52691b4dbfc