HTMLIFrameElement.sandbox should be a DOMSettableTokenList https://bugs.webkit.org/show_bug.cgi?id=150377
Reviewed by Ryosuke Niwa.

Source/WebCore:

HTMLIFrameElement.sandbox should be a DOMSettableTokenList as per the
HTML specification:
- https://html.spec.whatwg.org/multipage/embedded-content.html#htmliframeelement

Chrome and Firefox match the specification but Safari/WebKit was uding
a DOMString.

Test: fast/frames/sandbox-attribute.html

* html/HTMLIFrameElement.cpp:
(WebCore::HTMLIFrameElement::sandbox):
(WebCore::HTMLIFrameElement::parseAttribute):
* html/HTMLIFrameElement.h:
* html/HTMLIFrameElement.idl:

LayoutTests:

* fast/frames/sandbox-attribute-expected.txt: Added.
* fast/frames/sandbox-attribute.html: Added.
Add layout test to check that HTMLIframeElement.sandbox is now a
DOMSettableTokenList.

* fast/frames/sandboxed-iframe-navigation-allowed-expected.txt:
* fast/frames/sandboxed-iframe-navigation-allowed.html:
Update test that was using shouldBe() with HTMLIframeElement.sandbox.
shouldBe() does an === check but which now fails before
HTMLIframeElement.sandbox is now a DOMSettableTokenList that can be
stringified, rather than a String.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@191388 268f45cc-cd09-0410-ab3c-d52691b4dbfc