Web Inspector: srcset attributes should have hyperlinks to the resources https://bugs.webkit.org/show_bug.cgi?id=150409
Patch by Joseph Pecoraro <pecoraro@apple.com> on 2015-10-21
Reviewed by Timothy Hatcher.

The "srcset" attribute parsing is a simplification of:
https://html.spec.whatwg.org/multipage/embedded-content.html#attr-img-srcset

The "srcset" attribute value is expected to be in comma
separated groups. Within each group we always expect a link,
and an optional descriptor. We want to linkify the link.

* UserInterface/Views/CodeMirrorAdditions.js:
(tokenizeSrcSetString):
(extendedXMLToken):
When parsing srcset attribute, parse links as link tokens
so that they are linkified in resource content views.

* UserInterface/Views/DOMTreeElement.js:
(WebInspector.DOMTreeElement.prototype._buildAttributeDOM):
When creating the DOMTree DOM for srcset attributes,
create link elements for the link values of attribute value.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@191421 268f45cc-cd09-0410-ab3c-d52691b4dbfc