[GTK] [Stable] InstallMissingMediaPluginsPermissionRequest not defined when building without GStreamer https://bugs.webkit.org/show_bug.cgi?id=148607
Unreviewed, build fix.

Patch by Philip Chimento <philip.chimento@gmail.com> on 2015-10-26

* UIProcess/gstreamer/InstallMissingMediaPluginsPermissionRequest.h:
Add a stub for the InstallMissingMediaPluginsPermissionRequest class
for the case when video is disabled or not using GStreamer. It's OK
that this class does nothing, because the associated C API contains
an assert_not_reached if video is disabled. We just need to define
the symbol so that the function signature will compile.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@191618 268f45cc-cd09-0410-ab3c-d52691b4dbfc