Do not sanitize user input for input[type=url] https://bugs.webkit.org/show_bug.cgi?id=150346 <rdar://problem/23243240>
Patch by Keith Rollin <krollin@apple.com> on 2015-10-28
Reviewed by Brent Fulgham.

The Windows version of DumpRenderTree's eventSender.keyDown handler
simulates user-entry by decomposing the given character into a virtual
key code and a set of modifier keys. The code that determines if the
shift key should be simulated only worked if the character was an
upper case character. Extend this support to also include symbols that
(on the U.S. keyboard) require holding down the shift key.

This change allows fast/forms/input-user-input-sanitization.html
to succeed, which requires us to simulate the entering of '@' and ':'.

* DumpRenderTree/win/EventSender.cpp:
(keyDownCallback):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@191704 268f45cc-cd09-0410-ab3c-d52691b4dbfc