[JHBuild] Do not try to update the dependencies if jhbuild configuration hasn't changed https://bugs.webkit.org/show_bug.cgi?id=150665
Patch by Carlos Garcia Campos <cgarcia@igalia.com> on 2015-10-30
Reviewed by Csaba Osztrogonác.

We check it only to wipe the directory or not, but we are still
tryihng to update every module even when the configuration hasn't
changed. This is problematic because it makes the whole build fail
when any repository fails to checkout an already checked out
revision, because the server is down for example. This is even
more problematic for the bots that skip perfectly valid revisions.
It's still possible to force an update passing --force command
line argument.

* Scripts/update-webkit-libs-jhbuild:
(deleteJhbuildMd5): Delete the md5sum files if exist.
(runJhbuild): Return early if the md5sum files haven't
changed. Only write them after a successful build and delete them
if the build fails to ensure we don't skip next run.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@191784 268f45cc-cd09-0410-ab3c-d52691b4dbfc