Regression(r192038): Crash in ~SandboxExtension() https://bugs.webkit.org/show_bug.cgi?id=150898
Reviewed by Alex Christensen.

We are seeing a couple of crashes under ~SandboxExtension() on the bots
after ~SandboxExtension(). The reason is that the SandBoxExtension is
still in use (useCount > 0) when destroyed. This happens due to a bug
in setDefersLoading() that could cause the load to get started more
than once and thus the SandboxExtension to be consumed more than once.

* NetworkProcess/NetworkResourceLoader.cpp:
(WebKit::NetworkResourceLoader::setDefersLoading):
(WebKit::NetworkResourceLoader::consumeSandboxExtensions):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@192044 268f45cc-cd09-0410-ab3c-d52691b4dbfc