Crash running webaudio/panner-loop.html https://bugs.webkit.org/show_bug.cgi?id=150200 <rdar://problem/23136282>
Reviewed by Jer Noble.

Source/WebCore:

Test: webaudio/panner-loop.html

This is based on the changes in Blink r164822:
https://codereview.chromium.org/130003002

Avoid infinitely recursing on audio nodes by keeping track of which nodes we've already
visited.

* Modules/webaudio/PannerNode.cpp:
(WebCore::PannerNode::pullInputs): Pass set of visited nodes so we don't revisit
nodes we've already serviced.
(WebCore::PannerNode::notifyAudioSourcesConnectedToNode): Accept visitedNodes argument
so we can avoid revisiting nodes. Check if the current node has already been visited
before processing it.
* Modules/webaudio/PannerNode.h:

LayoutTests:

This is based on the changes in Blink r164822:
https://codereview.chromium.org/130003002

* webaudio/panner-loop-expected.txt: Added.
* webaudio/panner-loop.html: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@192281 268f45cc-cd09-0410-ab3c-d52691b4dbfc