ASSERTION FAILED: m_isEngaged in WTF::Optional::value https://bugs.webkit.org/show_bug.cgi?id=151094
Reviewed by Darin Adler.

Source/WebCore:

That ASSERT was hit because the precondition was incorrectly
met, i.e., we're considering that the main axis length was
definite when it was actually not. The problem is that to
determine whether it was indefinite or not we're using
RenderBox::hasDefiniteLogicalHeight() which did not perfectly
match RenderBox::computePercentageLogicalHeight() for the case
of RenderTables. So computePercentageLogicalHeight() was
returning Nullopt (i.e. indefinite) while
hasDefiniteLogicalHeight() was returning true (so definite).

Some checks were refactored in a helper function to solve the
inconsistency so that both functions behave the same way even
in this particular situation.

Test: css3/flexbox/inline-flex-percentage-height-in-table-crash.html

* rendering/RenderBox.cpp:
(WebCore::tableCellShouldHaveZeroInitialSize):
(WebCore::RenderBox::computePercentageLogicalHeight):
(WebCore::percentageLogicalHeightIsResolvable):
(WebCore::RenderBox::percentageLogicalHeightIsResolvableFromBlock):
* rendering/RenderBox.h:

LayoutTests:

* css3/flexbox/inline-flex-percentage-height-in-table-crash-expected.txt: Added.
* css3/flexbox/inline-flex-percentage-height-in-table-crash.html: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@192413 268f45cc-cd09-0410-ab3c-d52691b4dbfc