[GTK] Update xserver version in jhbuild https://bugs.webkit.org/show_bug.cgi?id=151383
Reviewed by Alejandro G. Castro.

In r192568 I assumed the bots would have a recent enough Xvfb
to use the -displayfd option, but I forgot we build our own
xserver just for Xvfb. The version we build is too old, so
we could just use an updated version (the same bots have
installed, for example) to use the -displayfd option.

* gtk/jhbuild.modules: Update xserver to 1.16.4.
* gtk/patches/xserver-fix-glx-init.patch: Deleted.
* gtk/patches/xserver-remove-bogus-dependencies.patch: Updated to
apply on 1.16.4.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@192572 268f45cc-cd09-0410-ab3c-d52691b4dbfc