REGRESSION(r192753): Remote Web Inspector: Enabling Remote Inspection on Auto Inspect candidate Debuggable doesn't show up in debuggers https://bugs.webkit.org/show_bug.cgi?id=151792
Patch by Joseph Pecoraro <pecoraro@apple.com> on 2015-12-03
Reviewed by Brian Burg.

* inspector/remote/RemoteInspector.mm:
(Inspector::RemoteInspector::updateAutomaticInspectionCandidate):
When m_debuggablesMap was split into both m_targetMap and m_listingMap
this particular case was missed in updating both the target and listing
when the target is updated. We should match RemoteInspector::updateTarget
and update the listing map as the debuggable may have changed to be
allowed to debug.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@193188 268f45cc-cd09-0410-ab3c-d52691b4dbfc