Allow JavaScript to iterate over plugins for local files https://bugs.webkit.org/show_bug.cgi?id=151783 <rdar://problem/23692113>
Source/WebCore:

Reviewed by Alexey Proskuryakov.

Test: http/tests/plugins/plugin-javascript-access.html
      plugins/plugin-javascript-access.html

* page/Page.cpp:
(WebCore::Page::showAllPlugins): True if we set the debugging flag to show
all plugins, or if the document's origin is from a local file.
* page/Page.h:

LayoutTests:

Modify the existing 'local file' plugin tests so that they don't turn on the debugging
flag to allow iterating over all plugins. Then create a new HTTP test (based on the
existing plugin-javascript-access test) to make sure that plugin iteration is blocked
for non-local-file access.

Reviewed by Alexey Proskuryakov.

* http/tests/plugins/plugin-javascript-access-expected.txt: Added.
* http/tests/plugins/plugin-javascript-access.html: Added.
* plugins/plugin-javascript-access.html: Remove unneeded 'internals.setShowAllPlugins' call.
* plugins/script-tests/navigator-mimeTypes-length.js: Ditto.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@193363 268f45cc-cd09-0410-ab3c-d52691b4dbfc