Consider still matching an address expression even if B3 has already assigned a Tmp to it https://bugs.webkit.org/show_bug.cgi?id=150777
Reviewed by Geoffrey Garen.

We need some heuristic for when an address should be computed as a separate instruction. It's
usually profitable to sink the address into the memory access. The previous heuristic meant that
the address would get separate instructions if it was in a separate block from the memory access.
This was messing up codegen of things like PutByVal out-of-bounds, where the address is computed
in one block and then used in another. I don't think that which block owns the address
computation should factor into any heuristic here, since it's so fragile: the compiler may lower
something by splitting blocks and we don't want this to ruin performance.

So, this replaces that heuristic with a more sensible one: the address computation gets its own
instruction if it has a lot of uses. In practice this means that we always sink the address
computation into the memory access.

* b3/B3LowerToAir.cpp:
(JSC::B3::Air::LowerToAir::effectiveAddr):



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@193941 268f45cc-cd09-0410-ab3c-d52691b4dbfc