_touchEventRegions should return regions in the view's coordinates https://bugs.webkit.org/show_bug.cgi?id=152189 -and corresponding- rdar://problem/23188605
Reviewed by Dan Bernstein.

Source/WebKit/mac:

The comment here was actually out of date. It claimed that touch rectangles 
are in the coordinate system of the document, but we had actually changed 
them to be in the view’s coordinate system in order to fix issues with 
handling touch events in UIWebView. But now we are going back to having the 
touch rectangles be in the document’s coordinate system, so we should fix the 
rtl bugs here by converting to view coordinates before handing the rects off 
to iOS WK1 clients.
* WebView/WebView.mm:
(-[WebView _touchEventRegions]):

LayoutTests:

* fast/events/touch/ios/touch-event-rtl-expected.txt: Added.
* fast/events/touch/ios/touch-event-rtl.html: Added.



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@193997 268f45cc-cd09-0410-ab3c-d52691b4dbfc