Frequent crash under -[WKPDFView web_setMinimumSize:]_block_invoke https://bugs.webkit.org/show_bug.cgi?id=152266 <rdar://problem/22092676>
Reviewed by Simon Fraser.

Source/WebKit2:

* UIProcess/ios/WKPDFView.mm:
(-[WKPDFView web_setMinimumSize:]):
We weren't retaining the UIScrollView, but depending on it surviving
into this block's invocation.

We don't need to do this asynchronously anymore because the referenced
bug is fixed, so make it happen synchronously. This also fixes the crash
because there's no opportunity for the scroll view to be released.

Tools:

* TestWebKitAPI/TestWebKitAPI.xcodeproj/project.pbxproj:
* TestWebKitAPI/Tests/WebKit2Cocoa/WKPDFViewResizeCrash.mm: Added.
Add a test that crashed before this change, and doesn't crash after.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@194060 268f45cc-cd09-0410-ab3c-d52691b4dbfc