Allow JavaScript to iterate over plugins for local SecurityOrigins https://bugs.webkit.org/show_bug.cgi?id=152489
Reviewed by Alexey Proskuryakov.

Source/WebCore:

Instead of calling 'isLocalFile' on the URL, we check the Document's SecurityOrigin. If
no origin exists, we construct a SecurityOrigin from the URL, and check whether it
satisfies 'isLocal'.

* page/Page.cpp:
(WebCore::Page::showAllPlugins): Revised to use SecurityOrigin.

Source/WebKit2:

The 'getWebVisiblePluginInfo' is filtering plugins, even when we want to list
all plugins. To avoid this, we check the Document's SecurityOrigin. If no such
origin exists, we construct a SecurityOrigin from the URL.
        
If the relevant SecurityOrigin satsifies 'isLocal', we show all plugins.

* WebProcess/WebCoreSupport/WebPlatformStrategies.cpp:
(WebKit::WebPlatformStrategies::getWebVisiblePluginInfo):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@194367 268f45cc-cd09-0410-ab3c-d52691b4dbfc