StringBuilder often creates two StringImpls. <https://webkit.org/b/152662>
Reviewed by Anders Carlsson.

If StringBuilder didn't manage to guess the exact final length of the
string being built, it would still keep the buffer around, and the final
build process would then return a substring into that buffer.

This effectively yielded two StringImpls instead of one, with the substring
retaining the longer buffer StringImpl.

This patch improves the situation by having StringImpl crop the buffer's
m_length field to the final built string length. That way we never have to
return a substring.

* wtf/text/StringBuilder.cpp:
(WTF::StringBuilder::reifyString):
* wtf/text/StringImpl.h:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@194510 268f45cc-cd09-0410-ab3c-d52691b4dbfc